feat/ Admin manage window

+++ mrt-app/src/components/cards/manager.tsx
@@ -1,15 +1,18 @@
 import { useState, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
+import { useNavigate, useParams } from 'react-router-dom';
 import CreateCard from '../cards/createCard';
-import { GetCardList } from './getCardList';
 import CardEntry from './cardEntry';
 import { hasSessionToken } from '../../auth/sessionTokenManager';
 
+export interface Card {
+    uuid: string;
+    balance: number;
+    tappedIn: boolean;
+    tappedInStation?: string;
+};
+
 export const CardManager = () => {
-    interface Card {
-        uuid: string;
-        balance: number;
-    };
+
 
     const [cards, setCards] = useState<Card[]>([]);
     const [cardAction, setCardAction] = useState('');
@@ -22,8 +25,10 @@ export const CardManager = () => {
     const [reload, triggerReload] = useState(false);
 
     const handleDelete = async (uuid: string) => {
-        await deleteCard(uuid);
-        triggerReload(!reload);
+        if (window.confirm('Are you sure you want to delete this card?')) {
+            await deleteCard(uuid);
+            triggerReload(!reload);
+          }
     };
 
     //Refreshes the list of cards.
@@ -41,19 +46,19 @@ export const CardManager = () => {
          {
             hasToken ? (
                 <div className="flex">
-                <aside className="w-64 h-screen bg-gray-800 text-white p-6 space-y-6">
-                    <h1 className="text-xl font-bold">Card Management</h1>
-                    <button
-                        onClick={() => setCardAction('create')}
-                        className="w-full bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">
-                        Create UUID
-                    </button>
-                    <button
-                        onClick={() => setCardAction('read')}
-                        className="w-full bg-blue-500 hover:bg-green-700 text-white font-bold py-2 px-4 rounded">
-                        Read UUID
-                    </button>
-                </aside>
+                    <aside className="w-64 h-screen bg-gray-800 text-white p-6 space-y-6">
+                        <h1 className="text-xl font-bold">Card Management</h1>
+                        <button
+                            onClick={() => setCardAction('create')}
+                            className="w-full bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">
+                            Create UUID
+                        </button>
+                        <button
+                            onClick={() => setCardAction('read')}
+                            className="w-full bg-blue-500 hover:bg-green-700 text-white font-bold py-2 px-4 rounded">
+                            Read UUID
+                        </button>
+                    </aside>
 
                 <main className="flex-grow p-6 overflow-auto h-screen">
                     <div className=''>                        {
@@ -63,8 +68,9 @@ export const CardManager = () => {
                                     <CardEntry
                                         key={index}
                                         uuid={card.uuid}
+                                        tappedIn={card.tappedIn}
                                         balance={card.balance}
-                                        handleUpdate={handleDelete}
+                                        handleDelete={handleDelete}
                                     />                                    
                                 </>
 
@@ -105,7 +111,7 @@ export const deleteCard = async (uuid: string) => {
 
 export const updateCard = async (uuid: string, balance: number) => {
     try {
-        const response = await fetch(`http://localhost:5000/cards/update?uuid=${uuid}`, {
+        const response = await fetch(`http://localhost:5000/manage/cards/update?uuid=${uuid}`, {
             method: 'PUT',
             headers: {
                 'Content-Type': 'application/json'
@@ -120,6 +126,24 @@ export const updateCard = async (uuid: string, balance: number) => {
     } catch (error) {
         console.log(error);
     }
+}
+
+
+export const GetCardList = async () => {   
+    try {
+        const response = await fetch('http://localhost:5000/cards/get', {
+            method: 'GET',
+            headers: {
+                'Content-Type': 'application/json'
+            },
+        });
+
+        const data = await response.json();
+        return data;
+    } catch (error) {
+        console.log(error);
+    }
+
 }
